+++ app/gui2/src/components/GraphEditor/NodeWidgetTree.vue
@@ -1,17 +1,17 @@
 <script setup lang="ts">
 import { Tree } from '@/generated/ast'
+import { ForcePort } from '@/providers/portInfo'
 import { provideWidgetTree } from '@/providers/widgetTree'
 import { useGraphStore } from '@/stores/graph'
 import { useTransitioning } from '@/util/animation'
 import type { AstExtended } from '@/util/ast'
-import { ForcePort } from '@/util/callTree'
 import { computed, toRef } from 'vue'
 import NodeWidget from './NodeWidget.vue'
 
 const props = defineProps<{ ast: AstExtended }>()
 const graph = useGraphStore()
 const rootPort = computed(() => {
-  return props.ast.isTree(Tree.Type.Ident) && !graph.db.isKnownFunction(props.ast.astId)
+  return props.ast.isTree(Tree.Type.Ident) && !graph.db.isKnownFunctionCall(props.ast.astId)
     ? new ForcePort(props.ast)
     : props.ast
 })
