+++ app/gui2/src/components/GraphEditor/widgets/WidgetPort.vue
@@ -2,13 +2,13 @@
 import NodeWidget from '@/components/GraphEditor/NodeWidget.vue'
 import { injectGraphNavigator } from '@/providers/graphNavigator'
 import { injectGraphSelection } from '@/providers/graphSelection'
-import { injectPortInfo, providePortInfo } from '@/providers/portInfo'
+import { ForcePort, injectPortInfo, providePortInfo } from '@/providers/portInfo'
 import { Score, defineWidget, widgetProps } from '@/providers/widgetRegistry'
 import { injectWidgetTree } from '@/providers/widgetTree'
 import { useGraphStore } from '@/stores/graph'
 import { useRaf } from '@/util/animation'
 import { Ast, AstExtended } from '@/util/ast'
-import { ArgumentAst, ArgumentPlaceholder, ForcePort } from '@/util/callTree'
+import { ArgumentAst, ArgumentPlaceholder } from '@/util/callTree'
 import { useResizeObserver } from '@/util/events'
 import { Rect } from '@/util/rect'
 import { uuidv4 } from 'lib0/random'
