+++ app/gui2/src/providers/portInfo.ts
@@ -1,5 +1,7 @@
+import type { AstExtended } from '@/util/ast'
 import { identity } from '@vueuse/core'
 import { createContextStore } from '.'
+import { GetUsageKey } from './widgetUsageInfo'
 
 interface PortInfo {
   portId: string
@@ -8,3 +10,16 @@ interface PortInfo {
 
 export { injectFn as injectPortInfo, provideFn as providePortInfo }
 const { provideFn, injectFn } = createContextStore('Port info', identity<PortInfo>)
+
+/**
+ * Widget input type that can be used to force a specific AST to be rendered as a port widget,
+ * even if it wouldn't normally be rendered as such.
+ */
+export class ForcePort {
+  constructor(public ast: AstExtended) {
+    if (ast instanceof ForcePort) throw new Error('ForcePort cannot be nested')
+  }
+  [GetUsageKey]() {
+    return this.ast
+  }
+}
